+++ src/components/SetPlayerName/SetPlayerName.css
@@ -1,0 +1,40 @@
+/* Center the form and style it */
+.set-player-name-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+  border-radius: 5px;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
+  width: 300px;
+  margin-top: 50px;
+  /* Adjust this value to move the form down */
+}
+
+/* Style for the input fields to make them bigger */
+.input-field {
+  width: 100%;
+  padding: 15px;
+  /* Increase padding for larger input size */
+  margin: 10px 0;
+  border: 2px solid #ccc;
+  border-radius: 5px;
+  font-size: 18px;
+  /* Increase font size for larger text */
+}
+
+/* Style for the submit button */
+.submit-button {
+  padding: 15px 15px;
+  /* Adjust padding for a larger button */
+  background-color: #ffffff;
+  color: rgb(0, 0, 0);
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+}
+
+.submit-button:hover {
+  background-color: #c6ccc7;
+}
